refac: Renamed hello_world.py to utils.py to avoid function/filename clash.
