+++ jutl/utilities/__init__.py
@@ -1,8 +1,8 @@
-# Import hello_world.py so module
-# classes and functions are usable at
-# 'from jutl import formatting' level.
-from .hello_world import *
+# Import utils.py so module
+# functions are usable at
+# 'from jutl import utilities' level.
+from .utils import *
 
 # Only show functions specified in
 # submodule files to the outside world.
-__all__ = hello_world.__all__
+__all__ = utils.__all__
